% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtmixEM.R
\name{mvtmixEM}
\alias{mvtmixEM}
\title{Train a t mixture model.}
\usage{
mvtmixEM(
  x,
  lambda0 = NULL,
  mu0 = NULL,
  sigma0 = NULL,
  nu0 = NULL,
  G = NULL,
  G_range = NULL,
  tolerance = 1e-08,
  max_iterations = 1000,
  optimize_nu = TRUE,
  EM_variant = "EM",
  delta_seq = (max_iterations:1)/max_iterations,
  initialization = "kmeans",
  sc = "BIC",
  stoppage = "aitken"
)
}
\arguments{
\item{x}{Matrix of data.}

\item{lambda0}{Initial EM value for lambda, the vector of mixing probabilities. Must sum to 1.}

\item{mu0}{Initial EM value for mu, the list of mean vectors.}

\item{sigma0}{Initial EM value for sigma, the list of covariance matricies.}

\item{nu0}{Initial EM value for nu, the vector of degrees of freedom. Must be greater than 0. If nu0 is specified and initialization != 'none' then values for nu will not be generated.}

\item{G}{Number of components. Must be an integer greater than 1. Must specify exactly one of G or G_range.}

\item{G_range}{Vector of G values to test (each value must be an integer greater than 1). The best value will be chosen via a selection criterion. Must specify exactly one of G or G_range.}

\item{tolerance:}{EM algorithm convergence tolerance.}

\item{max_iterations:}{EM algorithm max iterations.}

\item{optimize_nu:}{Logical value indicating whether to optimize for nu. If false, nu0 must be specified.}

\item{EM_variant:}{Variant of the EM algorithm to use for model fitting. Must be one of 'EM', 'SEM', 'SAEM'.}

\item{delta_seq:}{Vector of delta values for the SAEM variant of EM. Must be a decreasing sequence with each value <= 1.}

\item{initialization:}{Initialization method. Must be one of 'none', 'random', or 'kmeans'. If not 'none', overwrites lambda0, mu0, and sigma0 and, if nu0 has not specified, each element of nu0 is randomally sampled from Unif(0,100).}

\item{sc:}{Selection criteria to choose optimal number of components if G_range has been specified. Must be one of 'aic', 'bic', or 'icl'.}

\item{stoppage:}{Stoppage method for EM algorithm. Must be one of 'likelihood' (i.e. convergence is determined on the loglikelihood) or 'aitken' (i.e. convergence is determined on the aitkens acceleration values).}
}
\value{
Returns a list of class tMM with elements:
\itemize{
\item lambda - Vector of mixing probabilities.
\item mu - List of mean vectors.
\item sigma - List of covariance matricies.
\item nu - Vector of degrees of freedom.
\item G -  Number of components. If G_range is specified then G is the value in G_range that yields the optimal sc value.
\item DG - Number of parameters estimated.
\item AIC - AIC value.
\item BIC - BIC value.
\item ICL - ICL value.
\item loglik - Vector of the log likelihood values at each iteration.
\item logLG - Final log likelihood value.
\item initial - List of the lambda, mu, sigma and nu used to initialize EM.
}
}
\description{
Train a t mixture model on data x.
}
