% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtmixInitializeEM.R
\name{mvtmixInitializeEM}
\alias{mvtmixInitializeEM}
\title{Generate EM initial values for a t mixture model.}
\usage{
mvtmixInitializeEM(x, nu0 = NULL, G = NULL, initialization = "kmeans")
}
\arguments{
\item{x}{Matrix of data.}

\item{nu0}{Vector of degrees of freedom. Must be greater than 0. If nu0 is specified then values for nu will not be generated.}

\item{G}{Number of components. Must be greater than 1.}

\item{initialization}{Initialization method. Must be either 'none', random' or 'kmeans'. If initialization != 'none' and nu0 has not been specified, each element of nu0 is randomally sampled from Unif(0,100).}
}
\value{
Returns a list with elements:
\itemize{
\item lambda0 - Vector of mixing probabilities.
\item mu0 - List of mean vectors.
\item sigma0 - List of covariance matricies.
\item nu0 - Vector of degrees of freedom.
}
}
\description{
Generate EM initial values for a t mixture model with G components.
}
