% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtmixll.R
\name{mvtmixll}
\alias{mvtmixll}
\title{Complete data log likelihood of a t mixture model.}
\usage{
mvtmixll(x, lambda = NULL, mu = NULL, sigma = NULL, nu = NULL)
}
\arguments{
\item{x}{Matrix of data.}

\item{lambda}{Vector of mixing probabilities. Must sum to 1.}

\item{mu}{List of mean vectors.}

\item{sigma}{List of covariance matricies.}

\item{nu}{Vector of degrees of freedom. Must be greater than 0.}
}
\value{
The complete data loglikelihood (numeric).
}
\description{
Returns the complete data likelihood of a t mixture model with data x and parameters lambda, mu, sigma, nu.
}
