% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvtmix.R
\name{rmvtmix}
\alias{rmvtmix}
\title{Sample from a t mixture model.}
\usage{
rmvtmix(n = 1, lambda = NULL, mu = NULL, sigma = NULL, nu = NULL)
}
\arguments{
\item{n}{Number of points to sample.}

\item{lambda}{Vector of mixing probabilities. Must sum to 1.}

\item{mu}{List of mean vectors.}

\item{sigma}{List of covariance matricies.}

\item{nu}{Vector of degrees of freedom. Must be greater than 0.}
}
\value{
An n x p matrix in which each row is an independently generated realization from the specified t mixture model.
}
\description{
Sample from a t mixture model with parameters lambda, mu, sigma, nu.
}
